+++ RELEASE_NOTES.rst
@@ -10,7 +10,7 @@ Release Notes
 
 *sklearndf* 2 adds support for `scikit-learn 1.0 <https://scikit-learn.org/1.0>`_,
 adds data frame support for clusterers along with additional API enhancements and
-improvements, and is now subject to static type checking with :mod:`mypy`.
+improvements, and is now subject to static type checking with |mypy|.
 
 2.0.0
 ~~~~~